+++ com.unity.cinemachine/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Occasional precision issue when camera rotation is exactly 180 degress, causing roitational flickering.
 - Bugfix: Deceleration at the end of axis range was too aggressive.
 - Deoccluder accommodates camera radius in all modes.
+- StateDrivenCamera: child camera enabled status and priority are now taken into account when choosing the current active camera.
 
 
 ## [3.0.0-pre.7] - 2023-05-04